Conflicts:
	central-js2/client/index.html

+++ wf-central/src/main/java/org/wf/dp/dniprorada/model/document/DocumentAccessHandler.java
@@ -12,7 +12,7 @@ import org.wf.dp.dniprorada.model.DocumentAccess;
 @Component
 public interface DocumentAccessHandler {
 
-    DocumentAccessHandler setDocumentType(Integer docTypeID);
+    DocumentAccessHandler setDocumentType(Long docTypeID);
     DocumentAccessHandler setAccessCode  (String sCode_DocumentAccess);
     DocumentAccessHandler setPassword    (String sPass);
